fix(wear): make the phone-side bridge actually compile and bundle
Two failures from carrying the wear entry, both because I verified the watch's
Kotlin and the TypeScript but never the phone module's Kotlin or a web bundle.
The module's Android source only compiles as part of the app's Gradle build via
autolinking, so nothing local had ever built it.

Kotlin: the module used `AsyncFunction(name) Coroutine@{ }`, which is not an
Expo API — `Coroutine@` is just a label, so the non-suspend overload was
selected and every Task.await() was a compile error. The real API is
`AsyncFunction(name).SuspendBody<R> { }`. The explicit type argument is
load-bearing: SuspendBody's zero-arg and one-arg overloads are ambiguous when
the lambda declares no parameter list.

Also dropped expo-modules-core from peerDependencies. As a `*` peer, npm
installed a nested 2.5.0 copy beside the hoisted 3.0.29 the app actually uses,
and the nested copy's Gradle plugin was then compiled with mismatched Kotlin
metadata (2.1.0 read by a 1.9.0 compiler). The module gets core from the app's
autolinking config, so declaring it at all was wrong.

Bundling: use-wear-bridge imported the Android-only native module
unconditionally, so the web bundle resolved it and the desktop build died on the
module's uncompiled `main`. Split to use-wear-bridge.android.ts with a no-op
base, per the Metro platform-extension rule in CLAUDE.md. That keeps the module
out of every non-Android bundle rather than teaching the desktop build to build
an Android module it will never load — Wear OS does not pair with iOS, so there
is genuinely nothing to run off Android.

build:app-deps also now builds the module, which is what the F-Droid Android
build needs.

Verified locally this time:
  :getpaseo-expo-wear-bridge:compileDebugKotlin  -> BUILD SUCCESSFUL
  expo export --platform web  with the module's build/ DELETED (the exact
    condition that failed CI)                    -> Exported
  expo export --platform android                 -> Exported
  app typecheck, lint, 23 wear tests, 10 watch tests
