+++ packages/expo-wear-bridge/package.json
@@ -14,7 +14,6 @@
   "license": "MIT",
   "peerDependencies": {
     "expo": "*",
-    "expo-modules-core": "*",
     "react-native": "*"
   }
 }
